+++ backend/src/main/java/com/sdp/common/Trade.java
@@ -1,6 +1,7 @@
 package com.sdp.common;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
-public record Trade(String id, String symbol, Side side, double price, double quantity, Instant timestamp) {
+public record Trade(String id, String symbol, Side side, BigDecimal price, BigDecimal quantity, Instant timestamp) {
 }
